Disable hiddenapi access for AdServices tests
This test seems to pass fine without allowing access to
hidden APIs/implementation details, so remove the configuration
that allows it to access those.

This has the side-effect of making testApiVersion meaningful to keep
around, as it now verifies a "normal" APK can access the symbol. Add
some comments to make that clear.

Test: atest CtsAdServicesDeviceTestCases
Change-Id: I3a811d7fe4767d5f62a2b22c87272a3e4d1a3de8