Fixes event reporting measurement flacky test
Bug: 247696929
The test is flacky as the code delegates the work to another thread.
Therefore, the checks on the test could take longer to reach. Removing
these checks as they are not necessary, all the information we need to
verify if the kill switch is off can be seen on the same processing thread.

Test: atest AdServicesServiceCoreUnitTests
Change-Id: Ib9c1ae72c1177c7aea682b5d2c21a7d567e3fd82