Fixes flaky test on AttributionJobService
The class is now calling DebugReportingJobService, which calls Flags and
are not mocked. As the test code calls that runs code on a separate
thread, the error is not seen always, therefore it is flaky. The fix is
to mock DebugReportingJobService so the class just test what needs to be
tested. Error only seen when running code on the module
AdServicesServiceCoreUnitTests.

Test: atest AdServicesServiceCoreUnitTests
Change-Id: I9c19a18c786c55ee570bba30f909da86738a31ee