Implement logic for storing the storage metrics in memory
For sdk_version in java_library for sdksandbox-statsd, we have changed it to system_server_current because system_server_current allows to use module-lib which is needed for pulled atoms.

Also, we have used a jarjar rule for com.android.modules.utils.** so that only sdksandbox uses this library, and other modules do not call sdksandbox for this module.
Reference: go/android-mainline-apis#static-linking.

Bug: b/257952392
Test: atest SandboxMetricsInfoUnitTest
atest SdkSandboxManagerServiceUnitTest#testLoadSdk_computeSdkStorage

Change-Id: Ifa5016791713a8649b5250bcc6edd8de591426a0