Initialize sandbox before it can be used
Before the sandbox can be used to load SDKs, it needs to be initialized,
death recipients linked etc. However, previously, the sandbox service
binder object was being stored and could be used to load other SDKs
before initialization is complete. This CL ensures that the sandbox
service variable is set after initialization.

Additionally, ensure that the sandbox is initialized even after
restarts.

Bug: 259387335
Test: atest SdkSandboxManagerServiceUnitTest
Change-Id: I77f8d1ca8c6fd756c241b4260dbf1ef8d1668c0b