Fixing Beta 1 blocker. Notification was displayed multiple times for the EU users (daily).
Reason why it wasn't detected before. The unit tests was incorrectly
implemented - missing verify checks. Manaul CUJ were executed one by
one, meaning there was no chance to see whether the notification will be
displayed again. Additionally there was a limited amount of devices with
the EU simcard (1) and notification was tested in the isolation and on
happy path.

Test: atest + manual
Bug: 259524718
Change-Id: I16700cad30b906b6afa2adcf02b55714123e77e6