testscenario: Add support for JUnit Before and After
Add support for JUnit Before and After annotations.
Doing this by adding a new API methods to query for methods
with an annotation in the SDK.
Could have made a new API method that also invoked them instead
which would have saved on RPC calls but meant we would have had
more APIs doing similar jobs.
We don't need to worry about the performance between these calls
as much so this is a good trade off.

Also introduced flags for enabling these so that we can introduce
this behaviour without breaking the tests. This can be used for
introducing new features in the future.

Test: Tested webkit Rb tests with these changes
Bug: 259548800
Change-Id: I7a327943beacd757939ae0a534cdce235c25c54e