Fix flaky test for testRunAdSelectionMultipleCAsJSCachedSuccess
Moving cache.cleanup() out of the critical path of selectAds(), ensured a performance gain, but made the cache cleanUp and eventual-async event.
This made the test case flaky that asserted that cache cleanUp be run immediately after selectAds() run.

Fixed this by adding a wait in the mock verify

Bug: 260178718
Test: Ran unit test with 200 iterations and verified that its no longer
flaky
Change-Id: Ia9c19fa755cec2bf27af122da4456243966f8cc6

Change-Id: I2f688eda2ba3ccfa39086b8852ed5b32e509e4d8