Provide more specific package name for instrumented unit test
Out of three unit tests we have, only SdkSandboxFrameworkUnitTests test
is facing problem with Mockito MockMaker.

One of the difference I can see is that the package name of the
instrumented app overlaps with our actual code. Adding a "tests" suffix
to make it consistent with other tests we have.

Bug: 260314054
Test: atest SdkSandboxFrameworkUnitTests
Change-Id: I0ff69ce3c7b6f8a144648aa2c55d7f0b9506395b