Create test files under a subdirectory for better cleanup
The root directory of tests app's internal storge is also used by other
componenets. For example, mockitor's jar file is stored in test app's
cache directory which is under the root directory.

If we delete all files under root directory, we will end up deleting
these files too. We are only interested in deleting test files that we
created during the test. As such, a subdirectory under the root
directory is better for our usecase.

Bug: 260314054
Test: atest SdkSandboxFrameworkUnitTests
Change-Id: Ie92555f49c6b41d25ddfd896180d44b5e9c24086