Kill the app on sandbox death if needed
When the sandbox dies, the app should be killed if all of the following
conditions are satisfied:
* The app is alive.
* The app has atleast one SDK loaded.
* The app has not registered for sandbox death notifications (as
  otherwise it is assumed that the app wants to handle sandbox death).

Ideally, every app process that had loaded an SDK would die. However,
SdkSandboxManagerService does not currently distinguish between
different processes that load an SDK. As a result, currently only the
main process will be killed.

Bug: 261558210
Test: atest SdkSandboxLifecycleHostTest
Test: m statsd && statsd_testdrive 373 emits an atom with
SUBREASON_SDK_SANDBOX_DIED

Change-Id: Ifaa1c72f80e556b81cfa537159bb843d2fef2f8a