Support Database Downgrade
We added DB schema Flag to guard TopicContributorsTable:
http://cs/tm-mainline-prod/packages/modules/AdServices/adservices/service-core/java/com/android/adservices/data/DbHelper.java;l=177;rcl=86df023bb959d8b3162d91a180f0fea530517bb3

It's defaulted with disabled. When we enable it and kill adservices, db will be created with a higher version after calling adservices.

However, if we continue to disable it and kill adservices, db will throw an error of `not being able to downgrade database version` after calling adservices. This is because DB Helper doesn't implement onDowngrade method.

Bug: 261934022
Test: atest --iterations 50
Change-Id: Ice56a6826731adbd17e52353dbc9772a22284191