Add Mdd enrollment data file test.
The new test verifies Mdd enrollment data downloaded and processed successfully. We need to update this test when a new Mdd enrollment data is uploaded to the server.

Test: atest  AdServicesServiceCoreUnitTests AdServicesFrameworkUnitTests AdServicesEndToEndTests CtsAdServicesEndToEndTests CtsAdServicesDeviceTestCases AdServicesApkUnitTests CtsSandboxedTopicsManagerTests

Bug: 261918103
Change-Id: I6c5a2cbbe27c143016ea01a829915f2199a81658

Add Mdd enrollment data file test.

The new test verifies Mdd enrollment data downloaded and processed successfully. We need to update this test when a new Mdd enrollment data is uploaded to the server.

Test: atest  AdServicesServiceCoreUnitTests AdServicesFrameworkUnitTests AdServicesEndToEndTests CtsAdServicesEndToEndTests CtsAdServicesDeviceTestCases AdServicesApkUnitTests CtsSandboxedTopicsManagerTests

Bug: 261918103
Change-Id: I6c5a2cbbe27c143016ea01a829915f2199a81658