Create sdk-specific context to use as base for SandboxedSdkContext
Instead of using the sandbox context as base and customizing on top of
it, we now start with sdk's application info and customize from there
onwards.

For now, we modify two properties of sdk's application info:
  1. We modify the data path since sdks now have their own separate
  storage.
  2. We modify the package name so that permission checks do not break.

With the new customized context, most of the overriden methods of
SandboxedSdkContext is now redundant. We can remove them once we no
longer support T devices.

DD: go/sdk-context-redesign

Bug: 255937439
Bug: 263220446
Test: atest CustomizedSdkContextTest
Change-Id: I1c70018659b9fe011a086ed896fb2a99dc4706a2