Add instant_app CTS tests to sdksandbox-postsubmit TEST_MAPPING
This CL adds the CTS we wish to monitor within the sandbox to the
sdksandbox-postsubmit group. The list is taken from the current status
of go/cts-test-mapping-runs at the moment of writing.

Most of these tests do not pass. Going forward, we will want to look at
each of the tests in this list and decide whether we think that is
indeed a relevant test for the sandbox or not. If we decide it is
relevant, we will have to find a way to make the test pass. Otherwise,
we will remove the test from the list.

Note also that the test list contains only tests that are also enabled
for instant_app.

Testing this change:

Locally - single test
atest XYZTest -- \
        --enable-optional-parameterization \
	--enable-parameterized-modules \
	--module-parameter run_on_sdk_sandbox

Locally - test mapping group (with sharding)
atest --sharding 3 --test-timeout 60000 \
	--test-mapping packages/modules/AdServices/sdksandbox:sdksandbox-postsubmit -- \
	--enable-optional-parameterization \
	--enable-parameterized-modules \
	--module-parameter run_on_sdk_sandbox

Remotely on go/abtd-ui
Instructions: https://screenshot.googleplex.com/C39TXJXFktmeaU8
Results: https://android-build.googleplex.com/builds/abtd/run/L64500000958045539

Bug: 253447284
Test: see above
Change-Id: Ib292eb00273f7930569e122bafcf7e07ff8d6f7d