Add calling ApplicationThread to CallingInfo.
This allows us to use it later, when binding the SDK sandbox using the
new ActivityManagerLocal API. Note that currently,
callingApplicationThread can still be null inside CallingInfo, because
CallingInfo is constructed in many places where the calling application
thread isn't available.

The proper fix for this is probably to have a CallingInfo cache, that is
instantiated with a single instance of CallingInfo per caller, and
auto-created with the correct callingApplicationThread.

Bug: 253399592
Test: TEST_MAPPING
Change-Id: I4c25295e4d53eb89af5228c7cc97d94c77be5a66