LoadSdkSession: use calling user ID in createSdkProviderInfo
Currently createSdkProviderInfo uses getApplicationInfo to retrieve the
client's ApplicationInfo. If the client app is not installed on the
primary profile, then this call throws a NameNotFoundException.
Here we use getApplicationInfoAsUser and use the user ID corresponding
to the UID of the calling app. This allows the package to be found
regardless of calling UID.

We also add a CTS test to verify that when installed only for a
secondary user, the app can still load the SDK.

Fixes: 264257205
Test: atest SdkSandboxMultiUserHostTest
Change-Id: Ie7e94dde4a9d575e168cfc04f2817a3a3f865648
(cherry picked from commit 0e50962a3e2797432dbbf853df56ba78bbb8a741)
Merged-In: Ie7e94dde4a9d575e168cfc04f2817a3a3f865648