Configure feature flag for customized sdk context
We created a feature specific test: CustomizedSdkContextTest. Since all
the test cases in this suite need the feature to be enabled, we use adb
to turn on the feature for the suite.

For all other suite, we pass on the value from DeviceConfig. By default
the feature is disabled.

Bug: 255937439
Test: atest CustomizedSdkContextTest
Test: atest --test-mapping packages/modules/AdSservices/sdksandbox
Test: atest --test-mapping packages/modules/AdSservices/adservices
Test: turned on the feature using adb manually and ran all tests above
Change-Id: I01224c3411ce5866d37ae4760ec30b51f102c678
Merged-In: I01224c3411ce5866d37ae4760ec30b51f102c678