Clear consent for uninstalled app when packageUid = -1
Previous fix for this bug (ag/20941966) had cleared consent for all apps when packageUid = -1, rather than for all users of this app. This CL implements the desired behavior. The broadcast receiver looks for the package being removed in the set of installed packages, and assumes that the removal happened in a different user profile than the one it was associated with. If the package being removed was not present, it clears the consent data for that package.

Bug: 264587187
Test: atest PackageChangedReceiverTest AppConsentDaoTest BooleanFileDatastoreTest ConsentManagerTest
Change-Id: If958acb4926a26c95e8f38ce056bc62e8206ad70