Inform SdkSandboxService when loadSdk is successful
Previously, SdkSandboxService assumed that loading an SDK would
be successful by storing the loaded SDK name in mHeldSdk. This
caused consistency issues if the SDK failed to load, as
SdkSandboxManagerService and SdkSandboxService would have different
loading statuses. In this case, an SDK could not be reloaded in the
case that it initially failed to load.

This changes fixes the consistency issue by creating a new callback
between the SandboxedSdkHolder and SdkSandboxServiceImpl, that will
be invoked when the SDK has finished loading. When this occurs,
mHeldSdk will be updated accordingly. If the success callback is
not invoked, the SDK can be reloaded at a future time.

Test: atest SdkSandboxTest
Bug: 264532543
Change-Id: I2d10e86e8d89f266c49f152deb06c7fed861c846