Guard web trigger registration by XNA allow list
Details:
Parsing attribution config should not be allowed to non-allowlisted
packages for web converstions. Since the derived sources are created
using the trigger for XNA, a fraud app can register trigger with an
attribution-config which filters and uses sources from other AdTech
leading to attribution. We should avoid that happening.

Test: atest AdServicesServiceCoreUnitTests

On Pixel 4a -
https://screenshot.googleplex.com/AMphcthLXL5FRGC

Bug: b/266462242

Change-Id: I2230803566aae044bfbe6ca5ba52abd5906185f7