Start and stop the sandbox service
Start the sandbox service before binding to it. The BIND_AUTO_CREATE
flag no longer has to be used to bring up the sandbox. Additionally, the
service can return START_NOT_STICKY while starting to avoid being
restarted.

Additionally, stop the sandbox service instead of killing it by uid.
This is a step towards supporting shared uids if needed, and can also be
used to notify the sandbox when it is about to die.

Bug: 239562510
Bug: 230839879
Test: atest SdkSandboxLifecycleHostTest
Change-Id: I21b13ec360d7a9559dd8546848d91d4cc7e20990