Add canSendBroadcast API
The hook to check if sandbox process can send a broadcast currently
throws a SecurityException. The hook should return a boolean value and
ActivityManagerService should handle the denial of access before
throwing an exception.

The current method enforceAllowedToSendBroadcast is deprecated and a new method canSendBroadcast is being introduced here.

This was a feedback from API council to let platform to do internal cleanup and log metrics before throwing a SecurityException.

Bug: b/266069576

Test: atest SdkSandboxRestrictionsTest#testSendBroadcastsRestrictions_withAction
atest SdkSandboxRestrictionsTest#testSendBroadcastRestrictions_withoutAction

Change-Id: I20835f469b4f927025b490194c7b27813061439b