Add test to validate SandboxedSdkContextUtils handles SandboxedSdkContext correctly
Previously the test validated that instances of `Context` that were not instances of `SandboxedSdkContext` were handled correctly by the `getAsSandboxedSdkContext` method. However, it wasn't possible to test the handling of `SandboxedSdkContext` because merge conflicts between branches made it difficult to instantiate that object. Now that the merge conflicts have been resolved, add that test functionality.

Bug: 266693584
Fixed: 266693584
Test: atest SandboxedSdkContextUtilsTest
Change-Id: Ic9ca8bc3b5dfa0984084f66a1af8c8e79c153c86