Clear Blocked Topics in System Server when Topics Data is Reset
For example, consent is enabled and disabled. Narrow the scope to blocked topics only because we don't have an appropriate Manager to put this method in. And to put the method in in TopicsWorker requires to pass in AdServicesManager, which is a bit larger for current usage. Consider we won't preserve more data in near future, leave the refactoring to the next change.

Bug: 262274142
Test: atest
Change-Id: I54d648b3e76a8953d9603175602413ec96efd35b