Use package name instead of process name on R/S devices in CommonFixture
`CommonFixture` uses a call to `Process.myProcessName()` in the initializer of a static field. However, this API is only available on T+ devices. It appears that the process name by default is the same as the package name. Therefore, on R/S devices, use the Application context to fetch the package name in this static initializer instead.

Bug: 268499060
Test: atest AdServicesServiceCoreUnitTests AdServicesFrameworkUnitTests
Change-Id: I716ba007297cb95a0a50021eed6c2780fb0aac56