Set key as a string in aggregate contribution JSON
The change stores the number, which can be larger than 2^63, as a string
instead of an integer. The integer stores correctly in the datastore,
but when we use the JSONArray default parsing and apply
`JSONObject::getString` on that integer, which converts numbers greater
than 2^63 to scientific notation, the `AggregateCryptoConverter` can
fail to instantiate a `BigInteger`.

Bug: 268767829

Test: atest com.android.adservices.service.measurement

Change-Id: Ia685e98ae58c53a98490adfa86e772e6b88afc55