Add SdkSandboxManagerService#registerAppOwnedInterface
APIs to add and remove AppOwnedInterfaces in SdkSadboxManagerService
These APIs will be surfaced to an app developer to let them register implementations of interfaces. These implementations will be made available in the runtime for SDKs to use to invoke the App.

Bug: 267736592
Test: atest SdkSandboxManagerServiceUnitTest#testRegisteringAppOwnedInterfaceSuccess
atest SdkSandboxManagerServiceUnitTest#testRegisteringAppOwnedInterfaceAlreadyRegistered
atest SdkSandboxManagerServiceUnitTest#testUnRegisteringAppOwnedInterface
atest SdkSandboxManagerServiceUnitTest#testGetAppOwnedInterfaces

Change-Id: I85350b787e9572f02f73890c3c6d00265d3d08f1