Add SDK Sandbox UI Testing library
Creates a TestRule that handles device emulation,
setting up screenshot testing, loading an
SDK into the sandbox and starting an activity.
This can be used to verify that a view that is remotely
rendered from an SDK is rendered as expected. We use
an AlmostPerfectMatcher to account for slight rendering
differences between devices/architectures.

Individual test cases may interact with the running activity
to perform certain actions before verifying the rendered view.
Rendered views are compared with "golden" images that exist
in the assets directory.

Test: atest SdkSandboxUiTests
Bug: 268204038
Change-Id: I5b6ca4ef0eefe8da59910a73d4aa694d1b5563d8