Guard Fledge E2E tests behind a webview JS Sandbox availability check
The Fledge E2E tests are failing on R/S devices which have a system webview version earlier than the minimum required version. Once the system webview is updated to a newer version, the tests pass as-is. Adding this assumption that the JS Sandbox is available to the tests, so that they are skipped if running on a device that does not meet the minimum requirement.

Bug: 264462460
Bug: 269177610
Test: atest FledgeE2ETest
Change-Id: I9d2243dffcc846c08f9dadf42f6f616a7d417c18