Add SDKLevel check to guard client classes from invoking get() API.
Certain rollback tests seem to be failing due to the introduction of
back-compat get() API in ag/21324288. This is because the client classes
were altered to invoke the get() API but the test app is using the new
version of the client with PPAPIs rolled back, leading to failures
as get() API wouldn't exist in the rolled back version. As a workaround, we will be adding SDKLevel check to avoid invoking get() from the client classes on T+.

Bug: None
Test: lbi
Change-Id: Ibb38477bf5b2368c10bd45343a6766845a4cee2a