Fix Flaky PhFlagsTest
Currently, the order to get the value of some PhFlags is SystemProperties -> Device_Config -> DefaultVal. In CTS, for such PhFlags, we need to override the value via SystemProperties. However, SystemProperties doesn't support reset/remove a key so we can only revert the value back to the DefaultVal. In PhFlagsTest, we test the ability to override a flag through Device_Config. Therefore, such Flags will always return value from SystemProperties in this case.

To resolve the issue, added a new StaticMockFixture to allow to force returning the default value passed into SystemProperties.get() methods.

Bug: 266220611
Test: atest AdServicesServiceCoreUnitTests
Change-Id: I1eebb5e5a85f3c80d39582c9449919a8d1aa9f83