Revert the Classifier Type in TestPreparer
testTopicsManager_runOnDeviceClassifier overrides the classifier type to the on-device one. However, if this test fails, it will skip overriding the classifer type back to the original value at the end of this test. We need 1) explicity the classifier type in testTopicsManager_runDefaultClassifier, which is in the same test class. 2) In the TestPreparer, add a tear-down method to always revert the classifier type.

Bug: 271743989
Test: atest
Change-Id: Ia6cb3d76559f95bd2331083036976f22afcb54ad