Extends Waiting Time for Background Execution
Note the test is only flaky on aosp_cf_arm64_only_phone_hwasan-userdebug target. Extend this timeout aligning with other tests which use CountDownLatch. If this doesn't work, I'll convert the test to use CountDownLatch also.

The ABTD doesn't work for this target. Try to submit this change to see if it works.

Bug: 271670128
Test: atest
Change-Id: Iac5658de780e3a9fdb2e776bb1cfb3ea08e555bc