Fix the flaky test testAdServicesSystemService_enabled_then_disabled.
Issue: Flag value is changed above and TestableDeviceConfig invokes the DeviceConfig.OnPropertiesChangedListener. The listener is invoked on the separate threadi[1]. So, we need to add a wait time to ensure the listener gets executed. If listener gets executed after the test is finished, we hit READ_DEVICE_CONFIG exception.

After this fix, the test is passing 100% on T and U.

[1] https://source.corp.google.com/tm-mainline-prod/frameworks/libs/modules-utils/java/com/android/modules/utils/testing/TestableDeviceConfig.java;rcl=c8927342523b6a4440a4f9cf0ecb0d671bd35811;l=181

Test: atest AdServicesManagerServiceTest
http://ab/I57100010138445896

Bug: 271638773

Change-Id: Ie254f5c71783d3817dab9d0e3c485e6c609c9153