Fixes ad id test failing if rate limit reached
Bug: 271669337
The test normally doesn't fail, but fails if the device is acting
slowly. The dashboard shows the failure happened during one day and the
test was taking a little more than a second to complete. If the test
takes more than one second, the test is no longer reliable as the rate
limiter limits queries within a 1 second bound.

Test: atest CtsAdIdEndToEndTest:com.android.adservices.tests.adid.AdIdManagerTest#testAdIdManager_verifyRateLimitReached --iterations
Change-Id: I5ef17517f616248b40c7d246c02ec7329639a0b7