Extend CtsAdServicesPermissionsValidEndToEndTests to run on Android S.
Test file for S and T is mostly the same except for S we will need to
set some additional flags to avoid invoking systerm server related code.
The flags that needed to be set were identical to the ones from
CrystallBall testing. Hence, as part of this CL, created a more generic
CompatAdServicesTestUtils file that can be re-used across various tests.
Will clean up CompatTestUtils file as a follow-up.

Fix: 273530139
Test: atest CtsExtAdServicesPermissionsValidEndToEndTests
Change-Id: Ifa512dfa91f7c7f8ed5f543a19b545fc910f3524