Increase WIFI connection timeout to 20 seconds.
After we add the WIFI connection check in setup method, the test failure rate decreased on most of the targets. aosp_cf_arm64_only_phone_hwasan started to fail because it runs slow and needs more time to connect. I increase this timeout to 20 seconds. 20 seconds is a timeout limit, most of the tests only wait for less then 1 second.
https://android-build.corp.google.com/test_investigate/?testResultId=TR43728552867766067&invocationId=I16500010141464402

Test: atest AdServicesMddIntegrationTests

Bug: 272309378
Change-Id: I951b709b0a4e119d71be0025a4a9a4eeee55a9ff