Fixes tests that verify rate limit reached
Bug: 270671196
If the test takes more than one second, the test is no longer reliable as the rate limiter limits queries within a 1 second bound. The test could take longer than 1 second if there is slowness on the device. The test dashboard show that this is an intermittent issue, most of the times it is within the 1 second window.

Test: atest CtsAdServicesEndToEndTestMeasurement --iterations
Test: atest CtsAppSetIdEndToEndTest --iterations
Change-Id: Ieeaefca3de993b3c456093744d4c326628cace73