Implement input configuration for classifier (go/rb-ml-input-config)
This CL overhauls how classifier input is prepared: instead of
computing a fixed value from on-device data (currently, app
description), the structure and content of classifier input is
determined via a config file.

This file is bundled or downloaded along with other classifier assets
via ModelManager, which reads it into the new ClassifierInputConfig
class.  This class defines the format structure for the classifier input
along with the set of enumerated fields that will populate it.

ClassifierInputManager is now responsible for the entire process of
composing the classifier input; given a ClassifierInputConfig and a
package name, it will pull each required field value and format the
final input.

All input processing has been moved to this class, and
OnDeviceClassifier is agnostic to what input is passed in.
Correspondingly, PackageManagerUtil has been removed, as its
functionality has been folded into ClassifierInputManager.

The result is that the classifier model can now be updated without a
code change to take any combination of package name, short app name, or
manifest app description.  Models can be considered self-documenting in
this regard, telling the classifier what data to compose for their
input and how.

Bug: b/273584260
Test: atest ClassifierInputManagerTest CommonClassifierHelperTest
ModelManagerTest OnDeviceClassifierTest PrecomputedClassifierTest
Change-Id: I884bccb65e31784b3a001f06949df5809187a203

Change-Id: Idc45089dada8be06247c004ffdf96d9d9eae2359