Enable back-compat flag to fix Android S CB tests.
Android S CB tests for all APIs are failing in ATP. This is because
we created a new back-compat feature flag integrated to determine the
value of the global kill-switch. We need to enable the back-compat flag
in order for the global kill switch to be turned off on Android S.

Bug: 269498467
Test: m AdServicesScenarioTests -j $(nproc); ran loadtopics CB test
as an example. Will verify ATP runs directly for all other APIs.

Change-Id: If5391dd75157cf9f26a1950f6d247dfb988ff89e