Fix CtsExtAdServicesPermissionsValidEndToEndTests.
Certain setprop commands don’t seem to be running in post-submit. We’ll need to explicitly set these commands in the test file itself as a temporary workaround to unblock train boarding. We also had to set ppapi_app_signature_allow_list explicitly with a ‘*’ as a temporary for the test to pass. This is something that will require further investigation as well.

Bug: 274722195
Test: atest CtsExtAdServicesPermissionsValidEndToEndTests; abtd
Change-Id: I8b59ef486f9ca1f549b7bd36ec2be2c513be1c56