Consider sandbox importance for checks about app importance
This CL is updating CTS to start 2 sandbox activities, which should
cause 2 problems:
1- Failing on unloadSdk() at teardown, as we check if App in foreground,
which is false as the sandbox is in the forground (not the app) -> that
is fixed in this CL by considering the sandbox importance as well.
2- That would crash the system, as killing the test app after the test
is done would cause the sandbox activity to automaticall restart by
the system with no running sandbox process to host the activity ->
that is fixed in ag/22475342 by not restarting the sandbox activities.

MUST_SLEEP
Bug: 277052413
Bug: 277051501
Test: atest com.android.tests.sdksandbox.endtoend.SdkSandboxManagerTest

Change-Id: I467c95a90ce339b9233e9e1503c026f36a3c651e