Use unique android:authority strings for AdExtServices content providers.
We are noticing errors (see bug) in logs when attempting to record and trigger using measurement sample app when both AdServices and ExtServices are installed. There seems to be some resolution conflict in activating the right content provider due to identical android:authority strings being used for both AdServices and ExtServices. The fix is to make the android:authority strings unique.

Bug: b/277005735
Test: manual by following go/b277005735-debug multiple times; atest AsyncRegistrationContentProviderTest TriggerContentProviderTest
Change-Id: I3801aef4528269c77129a8b8f4eeb559065dc259