Handle restarting activities in test
In a lifecycle test, when the app process was killed after sandbox death, it was restarted automatically. This sometimes failed a check that ensured there was no app process in the process list. Instead, if the process restarts, compare the PIDs to make sure the old process was replaced by the new one.

Bug: 277731031
Change-Id: I1c4cd4aa16d45e3023636176a18e5a54810966b1
Test: atest SdkSandboxLifecycleHostTest