Ensure package and component package refer to sandbox package.
When apply sandbox activity checks to ensure that the sandbox process
can host it or at the moment of the sandbox activity interception, the
intent's package should be checked to exist and be equal to the sandbox
package to ensure that the intent-filter doesn't confuse the sandbox
activity with another activity with the same action.
It is also important to ensure that in case that intent's component
exist, its package should refer to the sandbox package name to ensure
that app do not start random activities in sandbox process, it is not
needed through to check compponent's class as package is unique already.

Bug: 274592326
Test: atest SdkSandboxManagerServiceUnitTest

Change-Id: I8ad9e0484c6e93cc0c2d1af204f2e4268eed0a6e