Use user-specific pm when retrieving info
Use the package manager instance for a particular user when getting the
provider information. This allows all calls (getApplicationInfo,
getProperty, and getPackageInfo) to be as the correct user.
We also use MATCH_ANY_USER when searching for SDK library as SDK can be installed for any user.

Bug: 272377545
Bug: 279932184
Test: atest --user-type secondary_user_on_secondary_display CtsSdkSandboxWebkitTestCases (on gcar_md_emu)
atest CtsSdkSandboxWebkitTestCases (on oriole)

Change-Id: Ib81e6dc8b646c7001e038b57c63b8eae007024e3