Add CTS test for AdServicesFrameworkInitializer
AdServicesFrameworkInitializer exposes a public method (registerServiceWrappers) that is visible to the System APIs and is supposed to be only invoked from SystemServiceRegistry during boot up. However, since this is a publicly exposed method, we need to add a CTS test to cover it. This method internally uses static methods on final classes that can't be mocked, and the expected behavior when invoked from anything other than SystemServiceRegistry is for it to throw an IllegalStateException. Therefore, we're adding a simple test that asserts the expected behavior.

Bug: 276946527
Test: atest CtsAdServicesDeviceTestCases
Change-Id: Ib6fa5fa61ef58c0640eec57159792a3222d71720