Remove SdkSandboxMultiUserHostTest
This test is not currently valid. It does not install the SDK
for the secondary user directly, meaning that once this fix ag/22690227
is merged the test will start failing.

We will need to re-write this test to install the SDK to secondary user
which will only be possible once b/271977755 is fixed.

Fixes: 267184124
Test: atest CtsSdkSandboxHostSideTests
Change-Id: I38c4e6dea7ae42bfc907a6d3559f56f50ca66b04