Fix FLEDGE CB tests to view showmap memory breakdown metrics.
Currently, on the FLEDGE CrystalBall dashboards we notice that showmap memory metrics are randomly missing for certain ATP runs (happens very often for Android S especially). This seems to be the case because there is a race condition between Showmap taking the snapshot at the end of each test and the execution of CleanPackageRule. Considering CleanPackageRule is not used in any of the PPAPI CB tests, I've disabled it from running at the end of each test (which is by default). This also aligns with the KillAppsRule, executed only before each test. Also moved back-compat flag setting into the SelectionAdsRule.

Bug: 269498467
Test: abtd tests all pass; able to view showmap metrics for the
AdServices process consistently.
https://android-build.googleplex.com/builds/abtd/run/L27300000960131020
https://android-build.googleplex.com/builds/abtd/run/L75700000960125082
https://android-build.googleplex.com/builds/abtd/run/L60000000960131327
https://android-build.googleplex.com/builds/abtd/run/L27600000960131822

Change-Id: I4fb5b3b2ef49a1b4f1380a8762a424c93820cb4f