SdkSandboxMetricsHostTest: Add tests for getHistoricalProcessExitReasons
Add test for SDK being able to access exit reasons for its own sandbox
process, and a test for an app with DUMP permission being able to access
exit reasons for the SDK sandbox package.

This is necessary because SDK sandbox has unique behaviour with respect
to real and package UIDs, therefore we need to ensure that we can index
using the real UID (for a single SDK sandbox) and package UID (for the
SDK sandbox package).

Fixes: 241524222
Test: atest SdkSandboxMetricsHostTest
Change-Id: I770d3323ef06b84f866175e38672afb3cefda9a4